Add files via upload
These files calculate properties related to composite materials, given the material's engineering constants with respect to the principal axes. In particular, 'Properties_multidirections.m' and 'Properties_principal2material' are used in the case of a UD lamina, while 'Eng_Const_Multilayer' obtains the engineering coefficients in the global reference of a multilayered composite (general case).